requirements.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
lxml==6.0.2
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
PyPDF2==3.0.1
pypdfium2==5.2.0
python-docx==1.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
